In the run worker, check the run status one more time after loading the pipeline definition
Summary:
This would account for the following race condition that a user reported;
- run worker monitoring is set up
- monitoring kicks in while the code is being loaded
- we blithely continue on, without checking if the run status is correc

The downside is one more  DB call on each run, but this is a pretty quick DB and/or API call I think?

Test Plan:
Testing the specific race condition reported here is tricky but can be done manually pretty easily by having a job that takes 3 minutes to load its repo, then marking it as failure in that period. Existing tests should also pass